Fix `outputs.templates` (#51)
Ref #45

The templates attribute in flake outputs must reference paths that point to a directory, not a file. This updates the available templates to provide starter projects that are easy to follow, mostly copied from the ipetkov/crane flake, but that include better descriptions and rust tooling that will make getting started much easier.